cloudant/requirements.txt
python-dotenv==1.0.0
ibm-cloud-sdk-core==3.10.0
ibmcloudant==0.0.34
cloudant==2.14.0

